Backend/pyproject.toml
[tool.poetry]
name = "areeb-competition-project"
version = "0.1.0"
description = ""
authors = ["OmarAlaraby <omar.alaraby23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"
cloudinary = "^1.44.0"
django-cors-headers = "^4.7.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
